+++ apps/web/app/bookings/[status]/page.tsx
@@ -1,0 +1,33 @@
+import { PageProps } from "app/_types";
+import { _generateMetadata } from "app/_utils";
+import { WithLayout } from "app/layoutHOC";
+import { redirect } from "next/navigation";
+import { z } from "zod";
+
+import { validStatuses } from "~/bookings/lib/validStatuses";
+import BookingsList from "~/bookings/views/bookings-listing-view";
+
+const querySchema = z.object({
+  status: z.enum(validStatuses),
+});
+
+export const generateMetadata = async () =>
+  await _generateMetadata(
+    (t) => t("bookings"),
+    (t) => t("bookings_description")
+  );
+
+export const generateStaticParams = async () => {
+  return validStatuses.map((status) => ({ status }));
+};
+
+const Page = async ({ params, searchParams }: PageProps) => {
+  const parsed = querySchema.safeParse({ ...params, ...searchParams });
+  if (!parsed.success) {
+    redirect("/bookings/upcoming");
+  }
+
+  return <BookingsList status={parsed.data.status} />;
+};
+
+export default WithLayout({ ServerPage: Page })<"P">;
