+++ turbo.json
@@ -242,7 +242,6 @@
     "APP_ROUTER_APPS_INSTALLED_CATEGORY_ENABLED",
     "APP_ROUTER_APPS_SLUG_ENABLED",
     "APP_ROUTER_APPS_SLUG_SETUP_ENABLED",
-    "APP_ROUTER_BOOKINGS_STATUS_ENABLED",
     "APP_ROUTER_EVENT_TYPES_ENABLED",
     "APP_ROUTER_AUTH_FORGOT_PASSWORD_ENABLED",
     "APP_ROUTER_AUTH_LOGIN_ENABLED",
